refactoring from code review
